+++ src/Frontend/Dfe.Complete/Pages/Projects/Notes/BaseProjectNotesModel.cs
@@ -10,7 +10,7 @@ using MediatR;
 
 namespace Dfe.Complete.Pages.Projects.Notes;
 
-public class BaseProjectNotesModel(ISender sender, string notesNavigation) : ProjectLayoutModel(sender, notesNavigation)
+public class BaseProjectNotesModel(ISender sender, ILogger _logger, string notesNavigation) : ProjectLayoutModel(sender, _logger, notesNavigation)
 {
     public async Task<NoteDto?> GetNoteById(Guid noteId)
     {