+++ CHANGELOG.md
@@ -31,8 +31,11 @@ Security in case of vulnerabilities.
 - Attach user ID from DB as custom claim
 
 ### Changed  
+- Add a redirect from projects/{id} to projects/{id}/tasks
+- Reduce log level from error to warning when project routes receive bad GUID or project not found 
 
 ### Fixed
+- Remove double pagination on Service Support > Local authorities
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-07-14.525...main) for everything awaiting release